[Nostromo] Refactor Composer Update, organize task tests better
