+++ src/Task/BuiltIn/Composer/DumpAutoloadTask.php
@@ -33,10 +33,10 @@ class DumpAutoloadTask extends AbstractTask
     public function execute()
     {
         $options = $this->getOptions();
-        $command = sprintf('%s dump-autoload %s', $options['path'], $options['flags']);
+        $cmd = sprintf('%s dump-autoload %s', $options['path'], $options['flags']);
 
         /** @var Process $process */
-        $process = $this->runtime->runCommand(trim($command));
+        $process = $this->runtime->runCommand(trim($cmd));
 
         return $process->isSuccessful();
     }
